% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{fil_dataset_uppdaterades}
\alias{fil_dataset_uppdaterades}
\title{När uppdaterades filen senast? (ISO 8601-sträng)}
\usage{
fil_dataset_uppdaterades(filnamn_med_sokvag)
}
\arguments{
\item{filnamn_med_sokvag}{Sökväg till filen.}
}
\value{
En sträng som \code{"2024-04-12T06:00:00Z"}.
}
\description{
När uppdaterades filen senast? (ISO 8601-sträng)
}
