% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{get_password_tk}
\alias{get_password_tk}
\title{Fråga efter ett lösenord i ett maskerat tcltk-fönster}
\usage{
get_password_tk(title = "Inloggning - lösenord", prompt = "Skriv lösenord:")
}
\arguments{
\item{title}{Fönstertitel.}

\item{prompt}{Ledtext.}
}
\value{
Lösenordet som sträng, eller \code{""} om användaren avbröt.
}
\description{
Fråga efter ett lösenord i ett maskerat tcltk-fönster
}
