% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{keyring_lagg_till_inloggning}
\alias{keyring_lagg_till_inloggning}
\title{Lägg till en databasinloggning i keyring}
\usage{
keyring_lagg_till_inloggning(keyring_service)
}
\arguments{
\item{keyring_service}{Namn på keyring-servicen, t.ex. \code{"databas_adm"}.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Frågar interaktivt efter användarnamn (via \code{svDialogs}) och lösenord (via
\code{\link[=get_password_tk]{get_password_tk()}}) och sparar dem för \code{keyring_service}. Finns en
inloggning redan frågas om den ska ersättas.
}
