% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felmeddelande.R
\name{logga_event}
\alias{logga_event}
\title{Logga en händelse till en loggfil}
\usage{
logga_event(meddelande, log_file)
}
\arguments{
\item{meddelande}{Text att logga.}

\item{log_file}{Sökväg till loggfilen (rader läggs till).}
}
\value{
Osynligt \code{NULL}.
}
\description{
Logga en händelse till en loggfil
}
