% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{postgres_alla_rattigheter_server}
\alias{postgres_alla_rattigheter_server}
\title{En användares rättigheter över alla databaser på servern}
\usage{
postgres_alla_rattigheter_server(
  anvandarnamn,
  con = "default",
  databaser = NULL,
  visa_databaser_utan_connect = TRUE,
  visa_alla_scheman = FALSE
)
}
\arguments{
\item{anvandarnamn}{Användare.}

\item{con}{En admin-\code{DBIConnection} eller \code{"default"} (använder
\code{\link[=uppkoppling_adm]{uppkoppling_adm()}}).}

\item{databaser}{Begränsa till dessa databaser (\code{NULL} = alla).}

\item{visa_databaser_utan_connect}{Ta med databaser användaren inte kan
ansluta till.}

\item{visa_alla_scheman}{\code{TRUE} = en rad per databas+schema; \code{FALSE} =
sammanräkning per databas och access-nivå.}
}
\value{
En \code{data.frame}.
}
\description{
En användares rättigheter över alla databaser på servern
}
