% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{postgres_anvandare_lagg_till}
\alias{postgres_anvandare_lagg_till}
\title{Lägg till en användare med lösenord}
\usage{
postgres_anvandare_lagg_till(con = "default", anvandarnamn, losenord)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{anvandarnamn, losenord}{Uppgifter för den nya användaren.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Lägg till en användare med lösenord
}
