% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{postgres_anvandare_ta_bort}
\alias{postgres_anvandare_ta_bort}
\title{Ta bort en användare/roll från servern}
\usage{
postgres_anvandare_ta_bort(con = "default", anvandarnamn)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{anvandarnamn}{Användare att ta bort.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Ta bort en användare/roll från servern
}
