% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databas.R
\name{postgres_databas_skapa}
\alias{postgres_databas_skapa}
\title{Skapa en ny databas}
\usage{
postgres_databas_skapa(con, databasnamn)
}
\arguments{
\item{con}{En aktiv \code{DBIConnection}.}

\item{databasnamn}{Namn på den nya databasen.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Skapa en ny databas
}
