% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databas.R
\name{postgres_databas_ta_bort}
\alias{postgres_databas_ta_bort}
\title{Ta bort en hel databas (kräver interaktiv bekräftelse)}
\usage{
postgres_databas_ta_bort(con, databasnamn)
}
\arguments{
\item{con}{En aktiv \code{DBIConnection}.}

\item{databasnamn}{Databas att ta bort.}
}
\value{
Osynligt \code{FALSE} om avbruten, annars \code{NULL}.
}
\description{
Ta bort en hel databas (kräver interaktiv bekräftelse)
}
