% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databas.R
\name{postgres_finns_schema_tabell_kolumner}
\alias{postgres_finns_schema_tabell_kolumner}
\title{Kontrollera att schema, tabell och kolumner finns}
\usage{
postgres_finns_schema_tabell_kolumner(
  con = "default",
  schema,
  tabell,
  kolumner = NULL,
  stoppa_vid_fel = TRUE
)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{schema, tabell}{Schema och tabell.}

\item{kolumner}{Teckenvektor med kolumnnamn (\code{NULL} = kontrollera inte).}

\item{stoppa_vid_fel}{\code{TRUE} = \code{stop()} vid fel; \code{FALSE} = returnera en lista
med \code{allt_finns} och \code{felmeddelande}.}
}
\value{
Osynligt \code{NULL} (om allt finns och \code{stoppa_vid_fel}), annars en lista.
}
\description{
Kontrollera att schema, tabell och kolumner finns
}
