% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grants.R
\name{postgres_grants_pa_befintliga_objekt}
\alias{postgres_grants_pa_befintliga_objekt}
\title{Ge läs-/skrivroller rättigheter på alla befintliga objekt}
\usage{
postgres_grants_pa_befintliga_objekt(
  con,
  lasroller = "lasroll",
  skrivroller = "skrivroll"
)
}
\arguments{
\item{con}{En aktiv \code{DBIConnection}.}

\item{lasroller, skrivroller}{Roller.}
}
\value{
Osynligt \code{TRUE}.
}
\description{
Ge läs-/skrivroller rättigheter på alla befintliga objekt
}
