% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista.R
\name{postgres_lista_behorighet_till_scheman}
\alias{postgres_lista_behorighet_till_scheman}
\title{Lista roller/användares behörighet per schema (read/write/no access)}
\usage{
postgres_lista_behorighet_till_scheman(con = "default")
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}
}
\value{
En \code{data.frame} med \code{role_or_user}, \code{table_schema}, \code{access_level}.
}
\description{
Lista roller/användares behörighet per schema (read/write/no access)
}
