% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista.R
\name{postgres_lista_databaser}
\alias{postgres_lista_databaser}
\title{Lista databaser på servern}
\usage{
postgres_lista_databaser(con = "default")
}
\arguments{
\item{con}{En \code{DBIConnection}, eller \code{"default"} för en tillfällig
standarduppkoppling.}
}
\value{
En \code{data.frame} med kolumnen \code{databas}.
}
\description{
Lista databaser på servern
}
