% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{postgres_lista_giltiga_rattigheter}
\alias{postgres_lista_giltiga_rattigheter}
\title{Giltiga PostgreSQL-rättigheter med beskrivning}
\usage{
postgres_lista_giltiga_rattigheter()
}
\value{
En \code{data.frame} med \code{Rattighet} och \code{Beskrivning}.
}
\description{
Giltiga PostgreSQL-rättigheter med beskrivning
}
