% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista.R
\name{postgres_lista_kolumnnamn_i_schema}
\alias{postgres_lista_kolumnnamn_i_schema}
\title{Lista kolumnnamn i alla tabeller i ett schema}
\usage{
postgres_lista_kolumnnamn_i_schema(con = "default", schema)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{schema}{Schemanamn.}
}
\value{
En \code{data.frame} med \code{table_name}, \code{column_name}, \code{data_type}.
}
\description{
Lista kolumnnamn i alla tabeller i ett schema
}
