% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista.R
\name{postgres_lista_roller_anvandare}
\alias{postgres_lista_roller_anvandare}
\title{Lista roller och användare}
\usage{
postgres_lista_roller_anvandare(con = "default")
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}
}
\value{
En \code{data.frame} från \code{pg_roles}.
}
\description{
Lista roller och användare
}
