% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista.R
\name{postgres_lista_rollmedlemskap}
\alias{postgres_lista_rollmedlemskap}
\title{Lista rollmedlemskap (vem ärver vilken roll)}
\usage{
postgres_lista_rollmedlemskap(con = "default")
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}
}
\value{
En \code{data.frame} med \code{user_or_role}, \code{inherited_role}.
}
\description{
Lista rollmedlemskap (vem ärver vilken roll)
}
