% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista.R
\name{postgres_lista_scheman_tabeller}
\alias{postgres_lista_scheman_tabeller}
\title{Lista scheman och deras tabeller/vyer}
\usage{
postgres_lista_scheman_tabeller(con = "default", visa_system_tabeller = FALSE)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{visa_system_tabeller}{Ta med \code{public}, \code{information_schema} och \verb{pg_*}.}
}
\value{
En namngiven lista: ett element per schema med en teckenvektor av
tabell-, vy- och matvy-namn.
}
\description{
Lista scheman och deras tabeller/vyer
}
