% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{postgres_losenord_byt_for_anvandare}
\alias{postgres_losenord_byt_for_anvandare}
\title{Byt lösenord för en användare}
\usage{
postgres_losenord_byt_for_anvandare(
  con = "default",
  anvandarnamn,
  nytt_losenord
)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{anvandarnamn, nytt_losenord}{Användare och nytt lösenord.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Byt lösenord för en användare
}
