% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{postgres_meta}
\alias{postgres_meta}
\title{Läs metadata-tabellen}
\usage{
postgres_meta(
  con = "default",
  tabell = "aktuell_version",
  schema = "metadata",
  query = NA
)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{tabell, schema}{Metadata-tabell och -schema.}

\item{query}{Valfri SQL-svans.}
}
\value{
En \code{data.frame}, eller \code{NULL} vid fel.
}
\description{
Läs metadata-tabellen
}
