% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{postgres_meta_skapa_vy_aktuell_version}
\alias{postgres_meta_skapa_vy_aktuell_version}
\title{Skapa vyn \code{metadata.aktuell_version}}
\usage{
postgres_meta_skapa_vy_aktuell_version(con)
}
\arguments{
\item{con}{En aktiv \code{DBIConnection}.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Vyn visar den senaste raden per schema/tabell ur \code{metadata.uppdateringar}.
}
