% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{postgres_metadata_uppdatera}
\alias{postgres_metadata_uppdatera}
\title{Lägg till en rad i metadata.uppdateringar}
\usage{
postgres_metadata_uppdatera(
  con,
  schema,
  tabell,
  version_datum = NA,
  version_tid = NA,
  uppdaterad_datum = Sys.Date(),
  uppdaterad_tid = format(Sys.time(), "\%H:\%M:\%S"),
  lyckad_uppdatering,
  kommentar = NA
)
}
\arguments{
\item{con}{En aktiv \code{DBIConnection}.}

\item{schema, tabell}{Vilket objekt raden gäller.}

\item{version_datum, version_tid}{Versionens datum/tid (\code{NA} = samma som
\verb{uppdaterad_*}).}

\item{uppdaterad_datum, uppdaterad_tid}{När uppdateringen gjordes.}

\item{lyckad_uppdatering}{\code{TRUE}/\code{FALSE}.}

\item{kommentar}{Fritextkommentar.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Skapar schemat \code{metadata} och tabellen \code{uppdateringar} om de saknas.
}
