% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{postgres_rattigheter_anvandare_lagg_till}
\alias{postgres_rattigheter_anvandare_lagg_till}
\title{Lägg till rättigheter för en användare i ett eller alla scheman}
\usage{
postgres_rattigheter_anvandare_lagg_till(
  con = "default",
  anvandarnamn,
  schema = "alla",
  rattigheter = c("CONNECT", "SELECT", "USAGE"),
  meddelande_rattigheter = TRUE
)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{anvandarnamn}{Användare.}

\item{schema}{\code{"alla"} eller en vektor av schemanamn.}

\item{rattigheter}{Vektor av rättigheter, eller \code{"alla"}. \code{c("CONNECT", "SELECT", "USAGE")} ger läsrättigheter.}

\item{meddelande_rattigheter}{Skriv ut varje beviljad rättighet.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Lägg till rättigheter för en användare i ett eller alla scheman
}
