% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{postgres_rattigheter_anvandare_ta_bort}
\alias{postgres_rattigheter_anvandare_ta_bort}
\title{Ta bort rättigheter från en användare}
\usage{
postgres_rattigheter_anvandare_ta_bort(
  con = "default",
  anvandarnamn,
  schema = "alla",
  rattigheter = "alla",
  meddelande_rattigheter = TRUE
)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{anvandarnamn}{Användare.}

\item{schema}{\code{"alla"} eller en vektor av schemanamn.}

\item{rattigheter}{Vektor av rättigheter, eller \code{"alla"}.}

\item{meddelande_rattigheter}{Skriv ut varje beviljad rättighet.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Ta bort rättigheter från en användare
}
