% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{postgres_roll_lagg_till}
\alias{postgres_roll_lagg_till}
\title{Skapa en roll (NOLOGIN) om den inte finns}
\usage{
postgres_roll_lagg_till(con = "default", rollnamn)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{rollnamn}{Rollnamn.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Skapa en roll (NOLOGIN) om den inte finns
}
