% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{postgres_roll_ta_bort}
\alias{postgres_roll_ta_bort}
\title{Ta bort en roll}
\usage{
postgres_roll_ta_bort(con = "default", rollnamn)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{rollnamn}{Rollnamn.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Ta bort en roll
}
