% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{postgres_roll_ta_bort_fran_anvandare}
\alias{postgres_roll_ta_bort_fran_anvandare}
\title{Ta bort en roll från en användare}
\usage{
postgres_roll_ta_bort_fran_anvandare(con = "default", rollnamn, anvandarnamn)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{rollnamn, anvandarnamn}{Roll och användare.}
}
\value{
Osynligt \code{TRUE}.
}
\description{
Ta bort en roll från en användare
}
