% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{postgres_roll_tilldela_till_anvandare}
\alias{postgres_roll_tilldela_till_anvandare}
\title{Tilldela en roll till en användare}
\usage{
postgres_roll_tilldela_till_anvandare(con = "default", rollnamn, anvandarnamn)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{rollnamn, anvandarnamn}{Roll och mottagande användare.}
}
\value{
Osynligt \code{TRUE}.
}
\description{
Tilldela en roll till en användare
}
