% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databas.R
\name{postgres_schema_finns}
\alias{postgres_schema_finns}
\title{Finns schemat?}
\usage{
postgres_schema_finns(con, schema_namn)
}
\arguments{
\item{con}{En aktiv \code{DBIConnection}.}

\item{schema_namn}{Schemanamn.}
}
\value{
\code{TRUE}/\code{FALSE}.
}
\description{
Finns schemat?
}
