% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databas.R
\name{postgres_schema_skapa_om_inte_finns}
\alias{postgres_schema_skapa_om_inte_finns}
\title{Skapa ett schema om det inte redan finns}
\usage{
postgres_schema_skapa_om_inte_finns(schema_namn, con = "default")
}
\arguments{
\item{schema_namn}{Schemanamn.}

\item{con}{En \code{DBIConnection} eller \code{"default"}.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Skapa ett schema om det inte redan finns
}
