% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databas.R
\name{postgres_schema_ta_bort}
\alias{postgres_schema_ta_bort}
\title{Ta bort ett tomt schema}
\usage{
postgres_schema_ta_bort(con, schema)
}
\arguments{
\item{con}{En aktiv \code{DBIConnection}.}

\item{schema}{Schemanamn.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Vägrar om schemat innehåller tabeller (skriver ut vilka).
}
