% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databas.R
\name{postgres_tabell_finns}
\alias{postgres_tabell_finns}
\title{Finns tabellen (eller vyn)?}
\usage{
postgres_tabell_finns(con = "default", schema, tabell)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{schema, tabell}{Schema och tabell.}
}
\value{
\code{TRUE}/\code{FALSE}.
}
\description{
Finns tabellen (eller vyn)?
}
