% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databas.R
\name{postgres_tabell_ta_bort}
\alias{postgres_tabell_ta_bort}
\title{Ta bort en tabell, vy eller materialiserad vy}
\usage{
postgres_tabell_ta_bort(con = "default", schema, tabell, drop_cascade = FALSE)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"}.}

\item{schema, tabell}{Schema och objektnamn.}

\item{drop_cascade}{Lägg till \code{CASCADE}.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Ta bort en tabell, vy eller materialiserad vy
}
