% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpostgres-package.R
\docType{package}
\name{rdpostgres-package}
\alias{rdpostgres}
\alias{rdpostgres-package}
\title{rdpostgres: PostgreSQL-verktyg för Region Dalarnas Databaser}
\description{
Anslutningar och administration mot Region Dalarnas PostgreSQL-servrar: lista och skapa databaser, scheman, roller och rättigheter, läsa och skriva tabeller, hantera metadata om när tabeller senast uppdaterades, samt automatiska grants via event triggers. Utbrutet ur postgres-avsnittet i func_GIS.R i Region-Dalarna/funktioner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Region-Dalarna/rdpaket}
  \item Report bugs at \url{https://github.com/Region-Dalarna/rdpaket/issues}
}

}
\author{
\strong{Maintainer}: Peter Möller \email{larspetermoller@gmail.com}

Authors:
\itemize{
  \item Peter Möller \email{larspetermoller@gmail.com}
}

Other contributors:
\itemize{
  \item Region Dalarna [copyright holder]
}

}
\keyword{internal}
