% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{uppdaterad_till_text_datum_tid}
\alias{uppdaterad_till_text_datum_tid}
\title{Dela upp en ISO 8601-uppdateringssträng i datum och tid}
\usage{
uppdaterad_till_text_datum_tid(datum_tid_txt)
}
\arguments{
\item{datum_tid_txt}{Sträng som \code{"2024-04-12T06:00:00Z"}.}
}
\value{
En lista med \code{datum} och \code{tid} (teckensträngar).
}
\description{
Dela upp en ISO 8601-uppdateringssträng i datum och tid
}
