% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppkoppling.R
\name{uppkoppling_adm}
\alias{uppkoppling_adm}
\title{Administrativ uppkoppling mot en PostgreSQL-databas}
\usage{
uppkoppling_adm(
  databas = "geodata",
  db_host = intern_db_host(),
  db_port = intern_db_port(),
  db_options = intern_db_options(),
  db_user = NA,
  db_password = NA
)
}
\arguments{
\item{databas}{Databasnamn.}

\item{db_host, db_port, db_options}{Anslutningsparametrar.}

\item{db_user, db_password}{Normalt via keyring; kan anges direkt.}
}
\value{
Ett \code{DBIConnection}-objekt, eller \code{NULL}.
}
\description{
Som \code{\link[=uppkoppling_db]{uppkoppling_db()}} men med keyring-service \code{databas_adm}.
}
